Migrate InitializationTests to addSourceLines
